Implement 최근검색뷰 using realm
