+++ MARU/MARU/Screens/Main/View/RecentSearchViewController.swift
@@ -46,22 +46,16 @@ final class RecentSearchViewController: BaseViewController {
     return deleteButton
   }()
 
-  private let emptyView = EmptytMeetingView()
   private let screenSize = UIScreen.main.bounds.size
   private var searchListCollectionView: UICollectionView! = nil
-  private var resultCollectionView: UICollectionView! = nil
   private var searchListDataSource: UICollectionViewDiffableDataSource<Section, String>!
-  private var resultDataSource: UICollectionViewDiffableDataSource<Section, BookModel>!
 
   private var viewModel =  RecentSearchViewModel()
-
-  private var initData = BookModel.initMainData
 
   override func viewDidLoad() {
     super.viewDidLoad()
     setNavigationBar(isHidden: false)
     configureLayout()
-    configureResultDataSource()
     bind()
   }
 
@@ -72,10 +66,6 @@ final class RecentSearchViewController: BaseViewController {
     configureSearchBar()
   }
 
-  @objc
-  private func didTapCancleButton() {
-    self.navigationController?.popViewController(animated: true)
-  }
   private func bind() {
     let viewWillAppear = rx.sentMessage(#selector(UIViewController.viewWillAppear(_: )))
       .map {_ in () }
@@ -92,32 +82,36 @@ final class RecentSearchViewController: BaseViewController {
     let output = viewModel.transform(input: input)
 
     output.load
-      .drive {
+      .drive { [self] in
         var recentKeywords: [String] = []
         $0.forEach { recentKeyword in
           recentKeywords.append(recentKeyword.keyword)
         }
-        self.configureSearchListDataSource(recentKeywords)
+        configureSearchListDataSource(recentKeywords)
       }
       .disposed(by: disposeBag)
 
     output.cancle
-      .drive {
+      .drive { [self] in
         if $0 {
-          self.view.hideKeyboard()
-          self.navigationController?.popViewController(animated: true)
+          view.hideKeyboard()
+          navigationController?.popViewController(animated: true)
         }
       }
       .disposed(by: disposeBag)
 
     output.delete
-      .drive(onNext: {
-        self.configureSearchListDataSource([])
+      .drive(onNext: { [self] in
+        configureSearchListDataSource([])
       })
       .disposed(by: disposeBag)
 
     output.keyword
-      .drive { print($0) }
+      .drive { [self] in
+        print($0)
+        let resultSearchViewController = ResultSearchViewController()
+        navigationController?.pushViewController(resultSearchViewController, animated: false)
+      }
       .disposed(by: disposeBag)
 
   }
@@ -145,16 +139,9 @@ extension RecentSearchViewController {
     searchListCollectionView = UICollectionView(frame: .zero,
                                                 collectionViewLayout: createListLayout())
     searchListCollectionView.delegate = self
-    resultCollectionView = UICollectionView(frame: .zero,
-                                            collectionViewLayout: MaruListCollectionViewLayout.createLayout())
     searchListCollectionView.contentInsetAdjustmentBehavior = .never
-    resultCollectionView.contentInsetAdjustmentBehavior = .never
-    resultCollectionView.delegate = self
     searchListCollectionView.backgroundColor = .white
-    resultCollectionView.backgroundColor = .white
-    resultCollectionView.isHidden = true
     searchListCollectionView.isHidden = false
-    emptyView.isHidden = true
 
     view.add(recentSearchLabel) {
       $0.snp.makeConstraints { make in
@@ -176,36 +163,6 @@ extension RecentSearchViewController {
         make.bottom.equalToSuperview()
       }
     }
-    view.add(resultCollectionView) {
-      $0.snp.makeConstraints { make in
-        make.top.equalTo(self.view.safeAreaLayoutGuide).inset(16)
-        make.leading.equalToSuperview()
-        make.trailing.equalToSuperview()
-        make.bottom.equalToSuperview()
-      }
-    }
-    view.add(emptyView) {
-      $0.snp.makeConstraints { make in
-        make.leading.equalToSuperview()
-        make.trailing.equalToSuperview()
-        make.height.equalTo(self.screenSize.height * 0.108)
-        make.centerY.equalTo(self.view.safeAreaLayoutGuide)
-      }
-    }
-  }
-
-  private func remakeLayout() {
-    searchListCollectionView.isHidden = true
-    deleteButton.isHidden = true
-    recentSearchLabel.isHidden = true
-    resultCollectionView.isHidden = false
-  }
-
-  private func backLayout() {
-    recentSearchLabel.isHidden = false
-    deleteButton.isHidden = false
-    searchListCollectionView.isHidden = false
-    resultCollectionView.isHidden = true
   }
 }
 
@@ -230,103 +187,9 @@ extension RecentSearchViewController {
     snapshot.appendItems(items)
     searchListDataSource.apply(snapshot, animatingDifferences: false)
   }
-
-  private func configureResultDataSource() {
-    let cellRegistration = UICollectionView.CellRegistration<MeetingListCell, BookModel> { (_, _, _) in
-      // Populate the cell with our item description.
-    }
-    resultDataSource
-      = UICollectionViewDiffableDataSource<Section, BookModel>(
-        collectionView: resultCollectionView
-      ) { (collectionView, indexPath, identifier ) -> UICollectionViewCell? in
-        return collectionView.dequeueConfiguredReusableCell(using: cellRegistration,
-                                                            for: indexPath,
-                                                            item: identifier)
-      }
-
-    var snapshot = NSDiffableDataSourceSnapshot<Section, BookModel>()
-    snapshot.appendSections([.main])
-    snapshot.appendItems(initData)
-    resultDataSource.apply(snapshot, animatingDifferences: false)
-  }
 }
 extension RecentSearchViewController: UICollectionViewDelegate {
   func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
     collectionView.deselectItem(at: indexPath, animated: true)
-  }
-}
-
-extension RecentSearchViewController: SearchTextFieldDelegate, UITextFieldDelegate {
-  func tapTextField() {
-    backLayout()
-  }
-}
-
-final class EmptytMeetingView: UIView {
-  private let emptyLabel: UILabel = {
-    let label = UILabel()
-    label.text =
-    """
-    이 책은 지금 개설된 모임이 없어요.
-    방장이 되어 직접 모임을 만들어보세요!
-    """
-    label.textAlignment = .center
-    label.font = .systemFont(ofSize: 13, weight: .medium)
-    label.textColor = .black
-    label.alpha = 0.22
-    label.numberOfLines = 2
-    return label
-  }()
-
-  private let openMeetingButton: UIButton = {
-    let button = UIButton()
-    button.contentEdgeInsets = UIEdgeInsets(top: 5,
-                                            left: 10,
-                                            bottom: 5,
-                                            right: 10)
-    let text = " 모임 열기"
-    let imageAttachment = NSTextAttachment()
-    imageAttachment.image = UIImage(systemName: "plus")?
-      .withTintColor(.mainBlue)
-      .withRenderingMode(.alwaysOriginal)
-
-    let multipleAttributes: [NSAttributedString.Key: Any] = [
-      .font: UIFont.systemFont(ofSize: 13, weight: .bold),
-      .foregroundColor: UIColor.mainBlue
-    ]
-
-    let attributeString =  NSMutableAttributedString(attachment: imageAttachment)
-    attributeString.append(NSAttributedString(string: text,
-                                              attributes: multipleAttributes))
-    button.setAttributedTitle(attributeString, for: .normal)
-
-    button.layer.borderWidth = 1
-    button.layer.borderColor = UIColor.mainBlue.cgColor
-    button.layer.cornerRadius = 13
-
-    return button
-  }()
-
-  override init(frame: CGRect) {
-    super.init(frame: frame)
-    configureLayout()
-  }
-
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-  private func configureLayout() {
-    add(emptyLabel) {
-      $0.snp.makeConstraints { make in
-        make.top.equalToSuperview()
-        make.centerX.equalToSuperview()
-      }
-    }
-    add(openMeetingButton) {
-      $0.snp.makeConstraints { make in
-        make.top.equalTo(self.emptyLabel.snp.bottom).inset(-13)
-        make.centerX.equalToSuperview()
-      }
-    }
   }
 }
