+++ MARU/MARU/Screens/Main/ViewModel/ResultSearchViewModel.swift
@@ -1,0 +1,11 @@
+//
+//  ResultSearchViewModel.swift
+//  MARU
+//
+//  Created by psychehose on 2021/07/28.
+//
+
+import Foundation
+
+import RxSwift
+import RxCocoa
